Fixes 2

+++ src/main/java/gateway/ClientRepository.java
@@ -23,12 +23,13 @@ public class ClientRepository implements ClientGateway {
     @Inject
     protected EntityManager em;
 
+    @Inject
+    private DatabaseService databaseService;
+
     @Override
     public Collection<Client> getClients() {
-        Collection <Client> clients = em.createQuery("SELECT c FROM Client c",
+        return em.createQuery("SELECT c FROM Client c",
             Client.class).getResultList();
-        
-        return clients;
     }
 
     @Override
@@ -71,14 +72,18 @@ public class ClientRepository implements ClientGateway {
     @Transactional
     public boolean createContact(String username, String email, String phone) {
         // getting client
-        Client client = this.getClientByName(username);
+        Client client = databaseService.getClientByName(username);
+
+        if (client.getContact() != null)
+            return this.updateContact(username, email, phone);
+
         // creating contact
         Contact contact = new Contact();
         contact.setEmail(email);
         contact.setPhone(phone);
-        contact.setClient(client);
         
         if (client != null){
+            contact.setClient(client);
             client.setContact(contact);
             em.merge(client);
             return true;
@@ -90,15 +95,14 @@ public class ClientRepository implements ClientGateway {
     @Transactional
     public boolean updateContact(String username, String email, String phone) {
         // getting client
-        Client client = this.getClientByName(username);
-        // creating contact
-        Contact contact = new Contact();
-        contact.setEmail(email);
-        contact.setPhone(phone);
-        contact.setClient(client);
+        Client client = databaseService.getClientByName(username);
+        
+        if (client.getContact() == null)
+            return false;
 
         if (client != null){
-            client.setContact(contact);
+            client.getContact().setEmail(email);
+            client.getContact().setPhone(phone);
             em.merge(client);
             return true;
         }
@@ -108,31 +112,21 @@ public class ClientRepository implements ClientGateway {
     @Override
     @Transactional
     public boolean deleteContact(String username) {
-        Client client = this.getClientByName(username);
+        Client client = databaseService.getClientByName(username);
+
+        if (client.getContact() == null)
+            return false;
 
         if (client != null){
+            client.getContact().deleteClient();
+            em.remove(client.getContact());
+
             client.deleteContact();
             em.merge(client);
+            
             return true;
         }
         return false;
-    }
-    
-    @Override
-    public Client getClientByName(String username) {
-        Client client = em.createQuery("Select c FROM Client c where " + 
-            "c.username LIKE :username",
-            Client.class)
-            .setParameter("username", username)
-            .getSingleResult();
-        
-        return client;
-    }
-
-    @Override
-    @Transactional
-    public void saveClient(Client client) {
-        em.merge(client);
     }
 
     private UserLogin createSecurityIdentityClient(String username, String password){
